#1363 Refactor test package of transaction.xa
