+++ sharding-transaction/sharding-transaction-xa/src/test/java/io/shardingsphere/transaction/xa/wrap/XADataSourceMapConverterTest.java
@@ -15,11 +15,10 @@
  * </p>
  */
 
-package io.shardingsphere.transaction.xa.manager;
+package io.shardingsphere.transaction.xa.wrap;
 
 import com.atomikos.jdbc.AtomikosDataSourceBean;
 import io.shardingsphere.core.constant.DatabaseType;
-import io.shardingsphere.transaction.xa.wrap.XADataSourceMapConverter;
 import org.apache.tomcat.dbcp.dbcp2.BasicDataSource;
 import org.junit.Test;
